split up avm tests
