+++ yarn-project/bb-prover/package.json
@@ -4,12 +4,9 @@
   "type": "module",
   "exports": {
     ".": "./dest/index.js",
-    "./wasm/bundle": "./dest/wasm/bundle.js",
-    "./wasm/lazy": "./dest/wasm/lazy.js",
     "./prover": "./dest/prover/index.js",
     "./verifier": "./dest/verifier/index.js",
-    "./test": "./dest/test/index.js",
-    "./config": "./dest/config.js"
+    "./test": "./dest/test/index.js"
   },
   "bin": {
     "bb-cli": "./dest/bb/index.js"
@@ -31,7 +28,7 @@
     "formatting": "run -T prettier --check ./src && run -T eslint ./src",
     "formatting:fix": "run -T eslint --fix ./src && run -T prettier -w ./src",
     "bb": "node --no-warnings ./dest/bb/index.js",
-    "test": "NODE_NO_WARNINGS=1 node --experimental-vm-modules ../node_modules/.bin/jest --passWithNoTests --maxWorkers=${JEST_MAX_WORKERS:-8}"
+    "test": "NODE_NO_WARNINGS=1 node --experimental-vm-modules ../node_modules/.bin/jest --passWithNoTests"
   },
   "jest": {
     "moduleNameMapper": {
@@ -59,15 +56,15 @@
       ".ts"
     ],
     "reporters": [
-      "default"
-    ],
-    "testTimeout": 30000,
-    "setupFiles": [
-      "../../foundation/src/jest/setup.mjs"
+      [
+        "default",
+        {
+          "summaryThreshold": 9999
+        }
+      ]
     ]
   },
   "dependencies": {
-    "@aztec/bb.js": "portal:../../barretenberg/ts",
     "@aztec/circuit-types": "workspace:^",
     "@aztec/circuits.js": "workspace:^",
     "@aztec/foundation": "workspace:^",
@@ -79,15 +76,12 @@
     "@noir-lang/noirc_abi": "portal:../../noir/packages/noirc_abi",
     "@noir-lang/types": "portal:../../noir/packages/types",
     "commander": "^12.1.0",
-    "pako": "^2.1.0",
     "source-map-support": "^0.5.21",
     "tslib": "^2.4.0"
   },
   "devDependencies": {
     "@aztec/ethereum": "workspace:^",
     "@aztec/kv-store": "workspace:^",
-    "@aztec/noir-contracts.js": "workspace:^",
-    "@aztec/protocol-contracts": "workspace:^",
     "@aztec/types": "workspace:^",
     "@jest/globals": "^29.5.0",
     "@types/jest": "^29.5.0",
@@ -98,7 +92,7 @@
     "jest-mock-extended": "^3.0.3",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4",
-    "viem": "2.22.8"
+    "viem": "^2.7.15"
   },
   "files": [
     "dest",
